backend/requirements.txt
fastapi>=0.102.0
uvicorn>=0.24.0
python-dotenv>=1.0.0
sqlalchemy>=2.0.0
alembic>=1.12.0
pytest>=7.0.0
httpx>=0.24.0
openai>=1.0.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.0.0,<5.0.0
email-validator>=2.0.0
